feat(export): return the model's UnitScales from the attribute export
`EntityRow` carries property and quantity values in the FILE's own units.
The geometry exporters normalise to metres; this path deliberately does
not, because a property value is not always a length and coercing one
would be guessing. That is a defensible choice, but it left the caller
needing the scale to interpret anything dimensional — and with no way to
obtain it, because `resolve_unit_scales` is internal to the processing
crate and `ExportModel` carried only entities.

The consequence is silent: a consumer writing quantities alongside
exported geometry gets a `Length` of 3000 from a millimetre model sitting
beside geometry in metres, with nothing in the API to detect it with.

So `stream_export_model` and `stream_export_model_with_index` now return
`UnitScales`, `ExportModel` carries a `units` field, and `UnitScales` is
re-exported from this crate.

Resolved from the `IFCPROJECT` id that pass 1 walks past, rather than
letting the resolver find it. Told which entity it is, the resolver does
an O(1) decode against the index this function already holds; not told,
it falls back to a SIMD substring search of the whole file. Pass 1 visits
every entity regardless, so recording the id costs nothing and no
exporter pays a scan it did not pay before.

The test is two synthetic files differing only in their declared length
unit. Their rows are byte-identical — which is the whole point — and
`units.length_unit_scale` is the only thing that tells them apart.

`UnitScales` gains `PartialEq` so `ExportModel` can keep deriving it.
Tests move to a sibling `model_tests.rs` under the house pattern: the
module-size ratchet counts non-test lines and `model.rs` was on its
budget. Budget lowered 655 → 530 to match.

Perf: `probe.sh AC20-FZK-Haus.ifc --iters 5` base vs branch — total 30 ms
vs 28, geometry 19 vs 18, index build 3.19 vs 3.03, all inside
run-to-run spread (base 30-35, branch 28-34). Output byte-identical: 285
meshes, 19,456 triangles, 35,940 vertices.
