+++ rust/processing/tests/module_size_allowlist.txt
@@ -27,7 +27,7 @@
     4028 rust/export/src/gltf.rs
 # COLLADA (.dae) exporter that embeds geometry in the KMZ (#1427); cohesive single-purpose module.
     635 rust/export/src/collada.rs
-     655 rust/export/src/model.rs
+     530 rust/export/src/model.rs
      611 rust/export/src/step.rs
 # +45: reject non-positive SegmentLength/HorizontalLength (malformed-input NaN guard) + regression test.
     1170 rust/geometry/src/alignment.rs
